add work for getting all users
--HG--
branch : noink

+++ src/noink/user_db.py
@@ -59,16 +59,27 @@ class UserDB:
         @return the user objects found
         '''
         users = [u]
-        if isinstance(u,IntType):
+        if isinstance(u, IntType):
             users = [self.find_user_by_id(u)]
         elif isinstance(u, string_types):
             users = self.find_user_by_name(u)
 
         return users
 
+    def get_all_users(self, g=None):
+        """
+        Get all users. Given an optional group, only find users in that
+        primary group.
+        """
+        if g is None:
+            return User.query.all()
+        else:
+            group = self.get_group(g)
+            return User.query.filter_by(primary_group=group).all()
+
     def get_user(self, u):
         '''
-         Given user information, returns the user object
+        Given user information, returns the user object
 
         @param u: The user to find. Can be an integer uid, username string, or
                   even a user object.
@@ -76,7 +87,7 @@ class UserDB:
         @return the user object found
         '''
         user = u
-        if isinstance(u,IntType):
+        if isinstance(u, IntType):
             user = self.find_user_by_id(u)
         elif isinstance(u, string_types):
             tu = self.find_user_by_name(u)
@@ -107,7 +118,8 @@ class UserDB:
             exists = False
 
         if exists:
-            raise DuplicateUser("%s already exists in database with id '%s'" % (username, str(exists)))
+            raise DuplicateUser("%s already exists in database with id '%s'".
+                format(username, str(exists)))
         else:
             from noink.role_db import RoleDB
             role_db = RoleDB()
@@ -135,7 +147,8 @@ class UserDB:
         Adds a new group to the database.
 
         @param group_name: The group name to add, must be unique.
-        @param user_id: (Optional) Single or multiple user IDs to associate with this group.
+        @param user_id: (Optional) Single or multiple user IDs to associate
+                        with this group.
 
         @return The group object created
         '''
@@ -143,7 +156,8 @@ class UserDB:
         exists = Group.query.filter_by(name=group_name).first()
 
         if exists:
-            raise DuplicateGroup("%s already exists in database with id '%s'" % (group_name, exists.id))
+            raise DuplicateGroup("%s already exists in database with id '%s'".
+                format(group_name, exists.id))
         else:
             g = Group(group_name)
             mainDB.session.add(g)
@@ -154,7 +168,9 @@ class UserDB:
                     gm = GroupMapping(g, exists)
                     mainDB.session.add(gm)
                 else:
-                    raise UserNotFound("%s not found in database to match with new group" % user_id)
+                    raise UserNotFound(
+                        "%s not found in database to match with new group".
+                        format(user_id))
 
             mainDB.session.commit()
             self.eventLog.add('add_group', 0, True, None, group_name)
@@ -172,7 +188,8 @@ class UserDB:
         user = self.get_user(u)
         group = self.get_group(g)
 
-        exist = GroupMapping.query.filter_by(user=user).filter_by(group=group).all()
+        exist = GroupMapping.query.filter_by(user=user).filter_by(
+                group=group).all()
 
         if exist == []:
             gm = GroupMapping(group, user)
@@ -217,7 +234,9 @@ class UserDB:
         gms = GroupMapping.query.filter_by(user=user)
 
         if gms is None:
-            raise UserHasNoGroups("%s does not have any group mappings! Every user should be a member of at least one group!" % user)
+            raise UserHasNoGroups(
+                "%s does not have any group mappings! Every user should be "\
+                "a member of at least one group!".format(user))
 
         groups = []
         for m in gms:
@@ -241,7 +260,8 @@ class UserDB:
         group = self.get_group(g)
 
         if isinstance(user, User) and isinstance(group, Group):
-            exists = GroupMapping.query.filter_by(user=user).filter_by(group=group).first()
+            exists = GroupMapping.query.filter_by(
+                    user=user).filter_by(group=group).first()
             if exists == []:
                 self.add_to_group(u, g)
             user.primary_group = group
@@ -263,7 +283,8 @@ class UserDB:
                     'Cannot remove {0} from their primary group {1}'.format(
                     user, group))
 
-            gm = GroupMapping.query.filter_by(user=user).filter_by(group=group).first()
+            gm = GroupMapping.query.filter_by(user=user).filter_by(
+                    group=group).first()
             if gm == []:
                 raise UserNotInGroup(
                     'User {0} was not a member of the group {1}'.format(
@@ -317,7 +338,8 @@ class UserDB:
         user = self.get_user(u)
         group = self.get_group(g)
 
-        exist = GroupMapping.query.filter_by(user=user).filter_by(group=group).first()
+        exist = GroupMapping.query.filter_by(user=user).filter_by(
+                group=group).first()
 
         if exist is None:
             return False
@@ -333,17 +355,18 @@ class UserDB:
         '''
 
         user = self.get_user(u)
-        # FIXME - Need a check here
-        uid = int(user.id)
-        uname = str(user.name)
-        mainDB.session.delete(user)
-        mainDB.session.commit()
-        self.eventLog.add('del_user', uid, True, None, uname)
+        if user is not None:
+            uid = int(user.id)
+            uname = str(user.name)
+            mainDB.session.delete(user)
+            mainDB.session.commit()
+            self.eventLog.add('del_user', uid, True, None, uname)
+        else:
+            raise UserNotFound("User not found in database")
 
     def authenticate(self, username, passwd, remember):
         '''
         Authenticates a user.
-        FIXME - docstring
         '''
         try:
             users = self.find_user_by_name(username)
@@ -360,20 +383,21 @@ class UserDB:
                 u.active = False
                 mainDB.session.commit()
                 return False
-        except: # FIXME - may want better handling
+        except: # XXX - may want better handling
             raise
-            return False
+        return False
 
     def logout(self, u):
         '''
-        Given a user, will log them out. Returns True on success, False on failure.
+        Given a user, will log them out. Returns True on success, False on
+        failure.
 
         @param u: A user to logout. Can be a uid or a user object.
         '''
         user = self.get_user(u)
-        if u.authenticated or u.active:
-            u.authenticated = False
-            u.active = False
+        if user.authenticated or user.active:
+            user.authenticated = False
+            user.active = False
             mainDB.session.commit()
             logout_user()
             return True
@@ -383,7 +407,6 @@ class UserDB:
 
 @loginManager.user_loader
 def _user_load(uid):
-    # FIXME - need try here?
     u = UserDB()
     return u.get_user(int(uid))
